Reorganize the CLI
Further reorganizes the CLI around Typer sub-commands and incorporates
MyPy  for the CLI commands.

Fixes PeopleForBikes/brokenspoke-analyzer/issues/247
