+++ brokenspoke_analyzer/cli/importer.py
@@ -1,17 +1,16 @@
-import asyncio
-
 import typer
 
 app = typer.Typer()
 
 
 @app.command()
-def all():
+def all() -> None:
     """Import all files into database."""
     # Retrieve the state info if needed.
-    asyncio.run(all_())
+    all_()
 
 
 # pylint: disable=too-many-arguments,duplicate-code
-def all_():
+def all_() -> None:
     """Import all files into database."""
+    return None
