+++ brokenspoke_analyzer/cli/root.py
@@ -1,8 +1,11 @@
 import asyncio
+import logging
 import pathlib
+import sys
 from typing import Optional
 
 import typer
+from loguru import logger
 from rich.console import Console
 
 from brokenspoke_analyzer.cli import (
@@ -16,6 +19,38 @@ from brokenspoke_analyzer.core import (
     runner,
 )
 
+
+def callback(verbose: int = typer.Option(0, "--verbose", "-v", count=True)) -> None:
+    """Define callback to configure global flags."""
+    # Configure the logger.
+
+    # Remove any predefined logger.
+    logger.remove()
+
+    # The log level gets adjusted by adding/removing `-v` flags:
+    #   None    : Initial log level is WARNING.
+    #   -v      : INFO
+    #   -vv     : DEBUG
+    #   -vvv    : TRACE
+    initial_log_level = logging.WARNING
+    log_format = (
+        "<level>{time:YYYY-MM-DDTHH:mm:ssZZ} {level:.3} {name}:{line} {message}</level>"
+    )
+    log_level = max(initial_log_level - verbose * 10, 0)
+
+    # Set the log colors.
+    logger.level("ERROR", color="<red><bold>")
+    logger.level("WARNING", color="<yellow>")
+    logger.level("SUCCESS", color="<green>")
+    logger.level("INFO", color="<cyan>")
+    logger.level("DEBUG", color="<blue>")
+    logger.level("TRACE", color="<magenta>")
+
+    # Add the logger.
+    logger.add(sys.stdout, format=log_format, level=log_level, colorize=True)
+
+
+# Create the CLI app.
 app = typer.Typer()
 app.add_typer(prepare.app, name="prepare", help="Prepare files needed for an analysis.")
 app.add_typer(importer.app, name="import", help="Import files into database.")
@@ -31,10 +66,18 @@ def compute(
     docker_image: Optional[str] = common.DockerImage,
     container_name: Optional[str] = common.ContainerName,
     city_fips: Optional[str] = common.CityFIPS,
-):
+) -> None:
     """Run an analysis."""
+    # Make MyPy happy.
+    if not docker_image:
+        raise ValueError("`docker_image` must be set")
+
     # Retrieve the state info if needed.
-    state_abbrev, state_fips = analysis.state_info(state) if state else (0, 0)
+    state_abbrev, state_fips = analysis.state_info(state) if state else ("0", "0")
+    if not state_abbrev:
+        raise ValueError("`state_abbrev` must be set")
+    if not state_fips:
+        raise ValueError("`state_fips` must be set")
     compute_(
         state_abbrev,
         state_fips,
@@ -49,15 +92,15 @@ def compute(
 
 # pylint: disable=too-many-arguments,duplicate-code
 def compute_(
-    state_abbrev,
-    state_fips,
-    city_shp,
-    pfb_osm_file,
-    output_dir,
-    docker_image,
-    container_name,
-    city_fips,
-):
+    state_abbrev: str,
+    state_fips: str,
+    city_shp: pathlib.Path,
+    pfb_osm_file: pathlib.Path,
+    output_dir: pathlib.Path,
+    docker_image: str,
+    container_name: str | None = None,
+    city_fips: str | None = None,
+) -> None:
     """Run the analysis."""
     console = Console()
     with console.status("[bold green]Running the full analysis (may take a while)..."):
@@ -87,8 +130,21 @@ def run(
     container_name: Optional[str] = common.ContainerName,
     city_fips: Optional[str] = common.CityFIPS,
     retries: Optional[int] = common.Retries,
-):
+) -> None:
     """Prepare all files and run an analysis."""
+    # Make MyPy happy.
+    if not docker_image:
+        raise ValueError("`docker_image` must be set")
+    if not speed_limit:
+        raise ValueError("`speed_limit` must be set")
+    if not block_size:
+        raise ValueError("`block_size` must be set")
+    if not block_population:
+        raise ValueError("`block_population` must be set")
+    if not retries:
+        raise ValueError("`retries` must be set")
+
+    # Prepare and run.
     asyncio.run(
         prepare_and_run(
             country,
@@ -108,18 +164,18 @@ def run(
 
 # pylint: disable=too-many-arguments
 async def prepare_and_run(
-    country,
-    state,
-    city,
-    output_dir,
-    docker_image,
-    speed_limit,
-    block_size,
-    block_population,
-    container_name,
-    city_fips,
-    retries,
-):
+    country: str,
+    state: str | None,
+    city: str,
+    output_dir: pathlib.Path,
+    docker_image: str,
+    speed_limit: int,
+    block_size: int,
+    block_population: int,
+    container_name: str | None,
+    city_fips: str | None,
+    retries: int,
+) -> None:
     """Prepare all files and run an analysis."""
     speed_file = output_dir / "city_fips_speed.csv"
     speed_file.unlink(missing_ok=True)
@@ -135,4 +191,4 @@ async def prepare_and_run(
         block_population,
         retries,
     )
-    compute.compute_(*params, docker_image, container_name, city_fips)
+    compute_(*params, docker_image, container_name, city_fips)
